new tkinter settings file
