+++ main.py
@@ -1,8 +1,7 @@
-from tkinter import *
-from tkinter import messagebox, ttk
 import pygame
 import json
 from backendStructure import *
+import tkSettingsWindow as settings
 
 pygame.init()
 
@@ -11,36 +10,12 @@ RES = (950, 700)
 clock = pygame.time.Clock()
 win = pygame.display.set_mode(RES)
 
-class Block():
-    components = []
-    def __init__(self, x, y, width, height, colour, text=""):
-        self.posX = x
-        self.posY = y
-        self.width = width
-        self.height = height
-        self.colour = colour
-        self.mainText = text
-        Block.components.append(self)
-    
-    @classmethod
-    def drawMenu(cls):
-        for com in cls.components:
-            pygame.draw.rect(win, com.colour, (com.posX, com.posY, com.width, com.height))
-            if len(com.mainText) > 0:
-                text = buttonFont.render(com.mainText, 1, colourScheme['buttonText'])
-                textPosX = int(RES[0] / 2 - text.get_width() / 2)
-                yOffset = 3
-                textPosY = int(com.posY + com.height/2 - text.get_height() / 2 + 3)
-                win.blit(text, (textPosX, textPosY))
 
-
-def ButtonSetup():
+def ButtonSetup(xPos, yPos, text):
     width = 300
     height = 50
-    xPos = int((RES[0] / 2) - (width / 2))
-    yPos = int(mainMenu.posY + round(mainMenu.height * 0.6))
 
-    return Block(xPos, yPos, width, height, colourScheme['button_bg'], text="Settings")
+    return Block(win, xPos, yPos, width, height, colourScheme['button_bg'], buttonFont, text=text)
 
 def buttonPressed(mPos):
     x, y = mPos
@@ -50,105 +25,6 @@ def buttonPressed(mPos):
     else: 
         return False
 
-def openSettings():
-
-    def addScale(*args):
-        global scalesDict, scaleNames, scale
-
-        name = newScaleName.get()
-        scaleString = newScaleNotes.get()
-
-        if not name=="" and not scaleString=="":
-            try:
-                scaleArray = [int(n) for n in scaleString.split(',')]
-                oldScales = readInScales()
-                oldScales[name] = scaleArray
-
-                with open('jsonScaleStorage.json', 'w') as file:
-                    json.dump(oldScales, file)
-
-                scalesDict = readInScales()
-                scaleNames = [s for s in scalesDict]
-                scale = StringVar(tkWindow, scaleNames[0])
-
-                scaleOptions.update()
-
-
-                
-            except:
-                print("That wasnt in the correct format deary")
-
-
-    def showHelp(*args):
-        messagebox.showinfo("Help", "To add a new scale, enter the name of the scale and the notes which make up the scale, with note #1 being the root note")
-
-    def showError(*args):
-        print("Not valid Scale")
-
-    def checkValid(*args):
-        if not newScaleNotes.get() == "":
-            try:
-                splitStringArray = [int(n) for n in newScaleNotes.get().split(',')]
-                if max(splitStringArray) >= 12 or min(splitStringArray) <= 0:
-                    # This is wrong -> ERROR
-                    showError()
-                else:
-                    # Correct -> Remove Error
-                    print("Thats good yeah")
-            except:
-                showError()
-
-    tkWindow = Tk()
-    tkWindow.geometry('290x230+500+500')
-    tkWindow.title("Scaley")
-
-    scalesDict = readInScales()
-    scaleNames = [s for s in scalesDict]
-    scale = StringVar(tkWindow, scaleNames[0])
-
-
-    note = StringVar(tkWindow, allNotes[0])
-
-    newScaleName = StringVar(tkWindow, "")
-    newScaleNotes = StringVar(tkWindow, "")
-    newScaleNotes.trace('w', checkValid)
-
-    Label(tkWindow, text="Choose Scale:").grid(row=0, column=0, padx=10)
-    Label(tkWindow, text="Root Note:").grid(row=1, column=0, padx=10)
-
-    ttk.Separator(tkWindow, orient='horizontal').place(x=0, y=85, relwidth=1)
-
-    Label(tkWindow, text="Add New Scale:").grid(row=3, columnspan=2, padx=10)
-
-    # Label(tkWindow, text="Show Note Letters: ").grid(row=3, column=0)
-
-    Label(tkWindow, text="Scale Name:").grid(row=4, column=0)
-    Label(tkWindow, text="Notes (eg: 1,3,4,5,6):").grid(row=4, column=1)
-
-
-    scaleOptions = OptionMenu(tkWindow, scale, *scaleNames)
-    scaleOptions.grid(row=0, column=1, padx=10)
-
-    noteOptions = OptionMenu(tkWindow, note, *allNotes)
-    noteOptions.grid(row=1, column=1, padx=10)
-
-    showLetter = IntVar() # IntVar is 0 or 1
-    showLetterToggle = Checkbutton(tkWindow, text="Show Note Letters: ", variable=showLetter).grid(row=2, columnspan=2)
-
-    scaleNameEntry = Entry(tkWindow, textvariable=newScaleName)
-    scaleNameEntry.grid(row=5, column=0, padx=10)
-
-    scaleNotesEntry = Entry(tkWindow, textvariable=newScaleNotes)
-    scaleNotesEntry.grid(row=5, column=1, padx=10)
-
-    Button(tkWindow, text="Add Scale", command=addScale).grid(row=7,column=0, pady=10)
-    Button(tkWindow, text="Help", command=showHelp).grid(row=7,column=1)
-    Button(tkWindow, text="Done", command=tkWindow.destroy).grid(row=8,columnspan=2)
-
-
-    mainloop()
-
-
 def drawFrame():
     win.fill(colourScheme['background'])
 
@@ -156,6 +32,7 @@ def drawFrame():
     myKeyboard.drawSharps()
     
     Block.drawMenu()
+    TextDisplay.drawAllText()
 
     pygame.display.update()
 
@@ -164,13 +41,18 @@ buttonFont = pygame.font.SysFont('Aldhabi', 40)
 titleFont = pygame.font.SysFont('Aldhabi', 40)
 textFont = pygame.font.SysFont('Aldhabi', 40)
 
+scaleHeading = TextDisplay(win, 130, 530, titleFont, "Current Scale")
+rootNoteHeading = TextDisplay (win, 600, 530, titleFont, "Current Root Note")
 
 myKeyboard = Keyboard(win, buttonFont)
-mainMenu = Block(0, 500, RES[0], RES[1] - 500, colourScheme['mainMenu'])
-settingsButton = ButtonSetup()
+mainMenu = Block(win, 0, 500, RES[0], RES[1] - 500, colourScheme['mainMenu'])
+settingsButton = ButtonSetup(80, 620, "Settings")
+controlsButton = ButtonSetup(570, 620, "Controls")
+
 
 CURRENT_ROOT_NOTE = "C"
 CURRENT_SCALE = "Major"
+
 
 running = True
 while running:
@@ -181,16 +63,17 @@ while running:
         elif event.type == pygame.MOUSEBUTTONDOWN:
             mousePos = pygame.mouse.get_pos()
             if buttonPressed(mousePos):
-                openSettings()
+                settings.openSettings()
         elif event.type == pygame.KEYDOWN:
             if allNotes.count(event.unicode.lower()) >= 1:
-                # myKeyboard.highlightKey(event.unicode)
                 print(CURRENT_ROOT_NOTE)
                 CURRENT_ROOT_NOTE = event.unicode
                 print(CURRENT_ROOT_NOTE)
                 myKeyboard.highlightScale(CURRENT_ROOT_NOTE, CURRENT_SCALE)
             elif event.unicode == 'p':
                 myKeyboard.highlightScale(CURRENT_ROOT_NOTE, CURRENT_SCALE)
+            elif event.key == pygame.K_SPACE:
+                myKeyboard.showScale = not myKeyboard.showScale                
     drawFrame()
         
 """
